+++ packages/terra-form-input/src/terra-dev-site/doc/common/BlankExample.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import Field from 'terra-form-field';
-import Input from '../../../Input';
+import Input from 'terra-form-input';
 
 const BlankExample = () => (
   <div>
     <Field label="Blank" htmlFor="blank">
-      <Input name="default blank input" id="blank" />
+      <Input name="default blank input" id="blank" ariaLabel="Blank" />
     </Field>
   </div>
 );