+++ packages/terra-form-input/src/terra-dev-site/doc/common/NumberInputExample.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import Field from 'terra-form-field';
-import Input from '../../../Input';
+import Input from 'terra-form-input';
 
 const NumberInputExample = () => (
   <div>
     <Field label="Numeric Input" htmlFor="numeric">
-      <Input name="number input" placeholder="enter digits" id="numeric" type="number" />
+      <Input name="number input" placeholder="enter digits" id="numeric" type="number" ariaLabel="Numeric Input" />
     </Field>
   </div>
 );