+++ packages/terra-form-input/src/terra-dev-site/test/form-input/RequiredFormInput.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Input from '../../../Input';
 
-const RequiredInput = () => <Input aria-label="Required testable form input" required id="form-input-required" />;
+const RequiredInput = () => <Input required id="form-input-required" ariaLabel="label" />;
 
 export default RequiredInput;