+++ packages/terra-form-input/tests/jest/Input.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Input from '../../src/Input';
 
 describe('Input', () => {
-  const defaultRender = <Input />;
+  const defaultRender = <Input ariaLabel="label" />;
 
   // Snapshot Tests
   it('should render a default component', () => {
@@ -11,25 +11,25 @@ describe('Input', () => {
   });
 
   it('should render as uncontrolled when just a default value is passed into the Input', () => {
-    const input = <Input defaultValue="foo" />;
+    const input = <Input defaultValue="foo" ariaLabel="label" />;
     const wrapper = shallow(input);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render with a type of password and pattern prop value', () => {
-    const input = <Input type="password" pattern=".{6,}" />;
+    const input = <Input type="password" pattern=".{6,}" ariaLabel="label" />;
     const wrapper = shallow(input);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render w/ placeholder text when just a placeholder prop is passed into the Input', () => {
-    const input = <Input placeholder="foo" />;
+    const input = <Input placeholder="foo" ariaLabel="label" />;
     const wrapper = shallow(input);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render as controlled when just a default value and `onChange()` is passed into the Input', () => {
-    const input = <Input value="foo" onChange={() => {}} />;
+    const input = <Input value="foo" onChange={() => { }} />;
     const wrapper = shallow(input);
     expect(wrapper).toMatchSnapshot();
   });
@@ -56,6 +56,18 @@ describe('Input', () => {
     const refCallback = jest.fn();
     const wrapper = mount(<Input refCallback={refCallback} />);
     expect(refCallback).toBeCalled();
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('should honor aria-label passed to component', () => {
+    const input = <Input defaultValue="foo" aria-label="label" />;
+    const wrapper = shallow(input);
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('should favor ariaLabel prop over aria-label if both props passed to component', () => {
+    const input = <Input defaultValue="foo" ariaLabel="ariaLabel" aria-label="aria-label" />;
+    const wrapper = shallow(input);
     expect(wrapper).toMatchSnapshot();
   });
 });