+++ packages/terra-form-radio/src/_RadioUtil.js
@@ -4,7 +4,7 @@ const isConsideredMobileDevice = () => window.matchMedia('(max-width: 1024px)').
       'ontouchstart' in window
       // eslint-disable-next-line no-undef
       || (window.DocumentTouch && document instanceof DocumentTouch)
-      || navigator.maxTouchPoints > 0
+      || navigator.maxTouchPoints > 0 // eslint-disable-line compat/compat
       || navigator.msMaxTouchPoints > 0
     );