+++ packages/terra-form-textarea/src/terra-dev-site/test/form-textarea/DefaultTextarea.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Textarea from '../../../Textarea';
 
-const textarea = () => <Textarea />;
+const textarea = () => <Textarea ariaLabel="label" />;
 
 export default textarea;