+++ packages/terra-image/src/terra-dev-site/doc/image/Image.1.doc.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import ImageSrc from '!raw-loader!../../../../src/Image';
-
 // Example Files
+import ImagePropsTable from '!terra-props-table-loader!../../../../src/Image';
+import ImageSrc from '!raw-loader!../../../../src/Image';
 import ImageDefault from '../example/ImageDefault';
 import ImageDefaultSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ImageDefault';
 import ImageFit from '../example/ImageFitTypes';
@@ -34,6 +34,7 @@ const DocPage = () => (
       {
         componentName: 'Image',
         componentSrc: ImageSrc,
+        componentProps: ImagePropsTable,
       },
     ]}
   />