+++ packages/terra-paginator/src/terra-dev-site/doc/paginator/ProgressivePaginator.3.doc.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/ProgressiveREADME.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import PaginatorSrc from '!raw-loader!../../../../src/ProgressivePaginator';
-
 // Example Files
+import ProgressivePaginatorPropsTable from '!terra-props-table-loader!../../../../src/ProgressivePaginator';
+import ProgressivePaginatorSrc from '!raw-loader!../../../../src/ProgressivePaginator';
 import ProgressivePaginatorExample from '../example/ProgressivePaginatorExample';
 import ProgressivePaginatorExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ProgressivePaginatorExample';
 
@@ -26,7 +26,8 @@ const DocPage = () => (
     propsTables={[
       {
         componentName: 'Progressive Paginator',
-        componentSrc: PaginatorSrc,
+        componentSrc: ProgressivePaginatorSrc,
+        componentProps: ProgressivePaginatorPropsTable,
       },
     ]}
   />