+++ packages/terra-show-hide/src/_ShowHideButton.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
-import KeyCode from 'keycode-js';
+import * as KeyCode from 'keycode-js';
 import styles from './_ShowHideButton.module.scss';